+++ HISTORY.rst
@@ -174,6 +174,7 @@
 - Two background tasks were added to make launching workflows an asynchronous
   event. This speeds up uploading documents in bulk.
 - Add the workflow action to the context of the initial state actions.
+- Add multiple workflow delete view.
 
 3.4.11 (2020-07-18)
 ===================